Refactor contacts controllers, add reexports, change in updateContact findOneAndUpdate to findOneAndReplace
