+++ controllers/contacts/index.js
@@ -1,0 +1,17 @@
+const getAllContacts = require('./getAllContacts');
+const getContactById = require('./getContactById');
+const addContact = require('./addContact');
+const deleteContactById = require('./deleteContactById');
+const updateContact = require('./updateContact');
+const updateFavorite = require('./updateFavorite');
+const checkAllowedMethods = require('./checkAllowedMethods');
+
+module.exports = {
+  getAllContacts,
+  getContactById,
+  addContact,
+  deleteContactById,
+  updateContact,
+  updateFavorite,
+  checkAllowedMethods,
+};
