+++ service/index.js
@@ -1,0 +1,5 @@
+const contactsService = require('./contacts-db');
+
+module.exports = {
+  contactsService,
+};
