--- utils/checkObjectID.js
@@ -1,12 +1,0 @@
-const ObjectId = require('mongoose').Types.ObjectId;
-
-// Validator function
-function isValidObjectId(id) {
-  if (ObjectId.isValid(id)) {
-    if (String(new ObjectId(id)) === id) return true;
-    return false;
-  }
-  return false;
-}
-
-module.exports = isValidObjectId;
